fix: proceed on advisory-only rectification exhaustion instead of failing green stories
decideStageAction failed a story whenever rectification exhausted with any
unfixed finding, even when every gate (tests/lint/typecheck/semantic/adversarial)
had passed and the only leftover was a sub-blocking advisory that no fix strategy
can claim (e.g. a `source:"autofix"` declaration diagnostic). Such a finding
forces a `no-strategy` cycle exit, which then hard-failed the story — observed on
nathapp-nestjs-platform event-bus-idempotency-dlq US-004.

- post-run.ts: add an advisory-only escape — if no remaining unfixed finding meets
  the run's blockingThreshold, return `continue`. Missing severity defaults to
  "error" (blocking) so real defects are never swallowed.
- cycle.ts: log the `no-strategy` exit at warn level with `orphanSources` so an
  unclaimable finding source is diagnosable instead of an invisible routing gap.
- tests: cover advisory-only → continue, blocking-present → escalate, missing
  severity → escalate, and stricter blockingThreshold. Split the oversized
  post-run-inspection test file by concern with a shared fixtures module; add
  makeTestContext to the @test/helpers barrel.
